\name{Barretts Data Package}
\alias{Barretts-package}
\alias{Barretts}
\docType{package}
\title{
Gene Expression profiles of normal esophogeal tissues, Barrett's esopogeal tissue and esopogeal adinocarcinoma.
}
\description{
This package contains only a data set which was downloaded from GEO. The dataset is described in "Progression of Barrett's metaplasia to adenocarcinoma is associated with the suppression of the transcriptional programs of epidermal differentiation."
Kimchi ET, Posner MC, Park JO, Darga TE, Kocherginsky M, Karrison T, Hart J, Smith KD, Mezhir JJ, Weichselbaum RR, Khodarev NN

}
\details{
\tabular{ll}{
Package: \tab Barretts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-02-20\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
This is just a data package. 
}
\author{
Nathan Morris
}
\references{
"Progression of Barrett's metaplasia to adenocarcinoma is associated with the suppression of the transcriptional programs of epidermal differentiation." Kimchi ET, Posner MC, Park JO, Darga TE, Kocherginsky M, Karrison T, Hart J, Smith KD, Mezhir JJ, Weichselbaum RR, Khodarev NN 
}
\examples{
data(barretts)
pData(barretts)
}
