\name{barretts}
\alias{barretts}
\docType{data}
\title{
Gene Expression profiles of normal esophogeal tissues, Barrett's esopogeal tissue and esopogeal adinocarcinoma.
}
\description{
This is an eSet with the data described in "Progression of Barrett's metaplasia to adenocarcinoma is associated with the suppression of the transcriptional programs of epidermal differentiation."
Kimchi ET, Posner MC, Park JO, Darga TE, Kocherginsky M, Karrison T, Hart J, Smith KD, Mezhir JJ, Weichselbaum RR, Khodarev NN

}
\usage{data(barretts)}
\format{
 Returns an eSet object
}

\source{
Progression of Barrett's metaplasia to adenocarcinoma is associated with the suppression of the transcriptional programs of epidermal differentiation."
Kimchi ET, Posner MC, Park JO, Darga TE, Kocherginsky M, Karrison T, Hart J, Smith KD, Mezhir JJ, Weichselbaum RR, Khodarev NN

}

\examples{
data(barretts)
pData(barretts)
}
\keyword{datasets}
barrett's esophogus